+++ src/Console/MutationCommand.php
@@ -8,7 +8,7 @@ class MutationCommand extends FieldGeneratorCommand
 {
     protected $name = 'lighthouse:mutation';
 
-    protected $description = 'Create a class for a single field on the root Mutation type.';
+    protected $description = 'Create a resolver class for a single field on the root Mutation type.';
 
     protected $type = RootType::MUTATION;