+++ src/Console/SubscriptionCommand.php
@@ -8,7 +8,7 @@ class SubscriptionCommand extends LighthouseGeneratorCommand
 {
     protected $name = 'lighthouse:subscription';
 
-    protected $description = 'Create a class for a single field on the root Subscription type.';
+    protected $description = 'Create a resolver class for a single field on the root Subscription type.';
 
     protected $type = RootType::SUBSCRIPTION;